Move register_(file|album)_*()aethods to new extension_points/event_hooks
